refacto component
